Vastly improved useful macro, and playing with values.
Disabled trails for now because they need to be fixed
